+++ examples/additional-test-suites/test/smoke/java/SayHelloSmokeTest.java
@@ -1,0 +1,22 @@
+import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.job.WorkflowJob;
+import org.jenkinsci.plugins.workflow.job.WorkflowRun;
+import org.junit.jupiter.api.Test;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.jvnet.hudson.test.junit.jupiter.WithJenkins;
+
+@WithJenkins
+class SayHelloSmokeTest {
+
+    @Test
+    void multipleCallsInOnePipeline(JenkinsRule jenkins) throws Exception {
+        WorkflowJob job = jenkins.createProject(WorkflowJob.class);
+        job.setDefinition(new CpsFlowDefinition(
+            "sayHello(); sayHello('Jenkins')",
+            true
+        ));
+        WorkflowRun run = jenkins.buildAndAssertSuccess(job);
+        jenkins.assertLogContains("Hello, world!", run);
+        jenkins.assertLogContains("Hello, Jenkins!", run);
+    }
+}