+++ examples/explicit-library-name/test/integration/java/GreetStepTest.java
@@ -1,0 +1,32 @@
+import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.job.WorkflowJob;
+import org.jenkinsci.plugins.workflow.job.WorkflowRun;
+import org.junit.jupiter.api.Test;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.jvnet.hudson.test.junit.jupiter.WithJenkins;
+
+@WithJenkins
+class GreetStepTest {
+
+    @Test
+    void defaultName(JenkinsRule jenkins) throws Exception {
+        WorkflowJob job = jenkins.createProject(WorkflowJob.class);
+        job.setDefinition(new CpsFlowDefinition("""
+                @Library('my-pipeline-lib') _
+                greet()
+                """, true));
+        WorkflowRun run = jenkins.buildAndAssertSuccess(job);
+        jenkins.assertLogContains("Hello, World!", run);
+    }
+
+    @Test
+    void customName(JenkinsRule jenkins) throws Exception {
+        WorkflowJob job = jenkins.createProject(WorkflowJob.class);
+        job.setDefinition(new CpsFlowDefinition("""
+                @Library('my-pipeline-lib') _
+                greet('Jenkins')
+                """, true));
+        WorkflowRun run = jenkins.buildAndAssertSuccess(job);
+        jenkins.assertLogContains("Hello, Jenkins!", run);
+    }
+}