+++ examples/explicit-library-name/vars/greet.groovy
@@ -1,0 +1,3 @@
+def call(String name = 'World') {
+    echo "Hello, ${name}!"
+}