+++ examples/library-resource/test/integration/java/LoadGreetingStepTest.java
@@ -1,0 +1,18 @@
+import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
+import org.jenkinsci.plugins.workflow.job.WorkflowJob;
+import org.jenkinsci.plugins.workflow.job.WorkflowRun;
+import org.junit.jupiter.api.Test;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.jvnet.hudson.test.junit.jupiter.WithJenkins;
+
+@WithJenkins
+class LoadGreetingStepTest {
+
+    @Test
+    void loadsAndEchoesGreeting(JenkinsRule jenkins) throws Exception {
+        WorkflowJob job = jenkins.createProject(WorkflowJob.class);
+        job.setDefinition(new CpsFlowDefinition("loadGreeting()", true));
+        WorkflowRun run = jenkins.buildAndAssertSuccess(job);
+        jenkins.assertLogContains("Hello from a library resource!", run);
+    }
+}